Refactor FeatureModel & FeatureNamespaceModel
