+++ featurebyte/api/feature_list.py
@@ -50,13 +50,13 @@ from featurebyte.enum import ConflictResolution
 from featurebyte.exception import RecordCreationException, RecordRetrievalException
 from featurebyte.feature_manager.model import ExtendedFeatureModel
 from featurebyte.models.base import PydanticObjectId, get_active_catalog_id
-from featurebyte.models.feature import DefaultVersionMode
 from featurebyte.models.feature_list import (
     FeatureListModel,
     FeatureListStatus,
     FeatureReadinessDistribution,
     FrozenFeatureListNamespaceModel,
 )
+from featurebyte.models.feature_namespace import DefaultVersionMode
 from featurebyte.models.tile import TileSpec
 from featurebyte.schema.deployment import DeploymentCreate
 from featurebyte.schema.feature_list import (
