+++ featurebyte/models/__init__.py
@@ -4,8 +4,9 @@ Document models for serialization to persistent storage
 from featurebyte.models.dimension_table import DimensionTableModel
 from featurebyte.models.entity import EntityModel
 from featurebyte.models.event_table import EventTableModel
-from featurebyte.models.feature import FeatureModel, FeatureNamespaceModel
+from featurebyte.models.feature import FeatureModel
 from featurebyte.models.feature_list import FeatureListModel
+from featurebyte.models.feature_namespace import FeatureNamespaceModel
 from featurebyte.models.feature_store import FeatureStoreModel
 
 all_models = [
