+++ featurebyte/models/feature.py
@@ -7,19 +7,18 @@ from typing import Any, List, Optional
 
 import pymongo
 from bson.objectid import ObjectId
-from pydantic import Field, PrivateAttr, StrictStr, root_validator, validator
+from pydantic import Field, PrivateAttr, root_validator, validator
 
-from featurebyte.common.doc_util import FBAutoDoc
-from featurebyte.common.validator import construct_sort_validator, version_validator
-from featurebyte.enum import DBVarType, OrderedStrEnum, StrEnum
+from featurebyte.common.validator import version_validator
+from featurebyte.enum import DBVarType
 from featurebyte.models.base import (
-    FeatureByteBaseModel,
     FeatureByteCatalogBaseDocumentModel,
     PydanticObjectId,
     UniqueConstraintResolutionSignature,
     UniqueValuesConstraint,
     VersionIdentifier,
 )
+from featurebyte.models.feature_namespace import FeatureReadiness
 from featurebyte.query_graph.graph import QueryGraph
 from featurebyte.query_graph.model.common_table import TabularSource
 from featurebyte.query_graph.model.graph import QueryGraphModel
@@ -27,177 +26,16 @@ from featurebyte.query_graph.node import Node
 from featurebyte.query_graph.node.metadata.operation import GroupOperationStructure
 
 
-class FeatureReadiness(OrderedStrEnum):
-    """Feature readiness"""
-
-    DEPRECATED = "DEPRECATED"
-    DRAFT = "DRAFT"
-    PUBLIC_DRAFT = "PUBLIC_DRAFT"
-    PRODUCTION_READY = "PRODUCTION_READY"
-
-
-class DefaultVersionMode(StrEnum):
+class BaseFeatureTargetModel(FeatureByteCatalogBaseDocumentModel):
     """
-    Default feature setting mode.
-    """
-
-    __fbautodoc__ = FBAutoDoc(proxy_class="featurebyte.DefaultVersionMode")
-
-    AUTO = "AUTO", "Automatically select the version to use."
-    MANUAL = "MANUAL", "Manually select the version to use."
-
-
-class FrozenFeatureNamespaceModel(FeatureByteCatalogBaseDocumentModel):
-    """
-    FrozenFeatureNamespaceModel store all the attributes that are fixed after object construction.
-    """
-
-    dtype: DBVarType = Field(
-        allow_mutation=False, description="database variable type for the feature"
-    )
-    entity_ids: List[PydanticObjectId] = Field(allow_mutation=False)
-    table_ids: List[PydanticObjectId] = Field(allow_mutation=False)
-
-    # pydantic validators
-    _sort_ids_validator = validator("entity_ids", "table_ids", allow_reuse=True)(
-        construct_sort_validator()
-    )
-
-    class Settings(FeatureByteCatalogBaseDocumentModel.Settings):
-        """
-        MongoDB settings
-        """
-
-        collection_name: str = "feature_namespace"
-        unique_constraints: List[UniqueValuesConstraint] = [
-            UniqueValuesConstraint(
-                fields=("_id",),
-                conflict_fields_signature={"id": ["_id"]},
-                resolution_signature=None,
-            ),
-            UniqueValuesConstraint(
-                fields=("name",),
-                conflict_fields_signature={"name": ["name"]},
-                resolution_signature=UniqueConstraintResolutionSignature.RENAME,
-            ),
-        ]
-
-        indexes = FeatureByteCatalogBaseDocumentModel.Settings.indexes + [
-            pymongo.operations.IndexModel("dtype"),
-            pymongo.operations.IndexModel("entity_ids"),
-            pymongo.operations.IndexModel("table_ids"),
-        ]
-
-
-class FeatureNamespaceModel(FrozenFeatureNamespaceModel):
-    """
-    Feature set with the same feature name
-
-    id: PydanticObjectId
-        Feature namespace id
-    name: str
-        Feature name
-    dtype: DBVarType
-        Variable type of the feature
-    feature_ids: List[PydanticObjectId]
-        List of feature version id
-    online_enabled_feature_ids: List[PydanticObjectId]
-        List of online enabled feature version id
-    readiness: FeatureReadiness
-        Aggregated readiness across all feature versions of the same feature namespace
-    created_at: datetime
-        Datetime when the FeatureNamespace was first saved or published
-    default_feature_id: PydanticObjectId
-        Default feature version id
-    default_version_mode: DefaultVersionMode
-        Default feature version mode
-    entity_ids: List[PydanticObjectId]
-        Entity IDs used by the feature
-    table_ids: List[PydanticObjectId]
-        Table IDs used by the feature
-    """
-
-    feature_ids: List[PydanticObjectId] = Field(allow_mutation=False)
-    online_enabled_feature_ids: List[PydanticObjectId] = Field(
-        allow_mutation=False, default_factory=list
-    )
-    readiness: FeatureReadiness = Field(allow_mutation=False)
-    default_feature_id: PydanticObjectId = Field(allow_mutation=False)
-    default_version_mode: DefaultVersionMode = Field(
-        default=DefaultVersionMode.AUTO, allow_mutation=False
-    )
-
-    # pydantic validators
-    _sort_feature_ids_validator = validator("feature_ids", allow_reuse=True)(
-        construct_sort_validator()
-    )
-
-    class Settings(FrozenFeatureNamespaceModel.Settings):
-        """
-        MongoDB settings
-        """
-
-        indexes = FrozenFeatureNamespaceModel.Settings.indexes + [
-            pymongo.operations.IndexModel("feature_ids"),
-            pymongo.operations.IndexModel("online_enabled_feature_ids"),
-            pymongo.operations.IndexModel("readiness"),
-            pymongo.operations.IndexModel("default_feature_id"),
-            pymongo.operations.IndexModel(
-                [
-                    ("name", pymongo.TEXT),
-                ],
-            ),
-        ]
-
-
-class FeatureModel(FeatureByteCatalogBaseDocumentModel):
-    """
-    Model for Feature asset
-
-    id: PydanticObjectId
-        Feature id of the object
-    name: str
-        Feature name
-    dtype: DBVarType
-        Variable type of the feature
-    graph: QueryGraph
-        Graph contains steps of transformation to generate the feature
-    node_name: str
-        Node name of the graph which represent the feature
-    tabular_source: TabularSource
-        Tabular source used to construct this feature
-    readiness: FeatureReadiness
-        Feature readiness
-    version: VersionIdentifier
-        Feature version
-    online_enabled: bool
-        Whether to make this feature version online enabled
-    definition: str
-        Feature definition
-    entity_ids: List[PydanticObjectId]
-        Entity IDs used by the feature
-    table_ids: List[PydanticObjectId]
-        Table IDs used by the feature
-    primary_table_ids: Optional[List[PydanticObjectId]]
-        Primary table IDs of the feature (auto-derive from graph)
-    feature_namespace_id: PydanticObjectId
-        Feature namespace id of the object
-    feature_list_ids: List[PydanticObjectId]
-        FeatureList versions which use this feature version
-    deployed_feature_list_ids: List[PydanticObjectId]
-        Deployed FeatureList versions which use this feature version
-    created_at: Optional[datetime]
-        Datetime when the Feature was first saved
-    updated_at: Optional[datetime]
-        When the Feature get updated
+    BaseFeatureTargetModel is the base class for FeatureModel & TargetModel.
+    It contains all the attributes that are shared between FeatureModel & TargetModel.
     """
 
     dtype: DBVarType = Field(allow_mutation=False, default=DBVarType.UNKNOWN)
     node_name: str
     tabular_source: TabularSource = Field(allow_mutation=False)
-    readiness: FeatureReadiness = Field(allow_mutation=False, default=FeatureReadiness.DRAFT)
     version: VersionIdentifier = Field(allow_mutation=False, default=None)
-    online_enabled: bool = Field(allow_mutation=False, default=False)
     definition: Optional[str] = Field(allow_mutation=False, default=None)
 
     # special handling for those attributes that are expensive to deserialize
@@ -205,15 +43,10 @@ class FeatureModel(FeatureByteCatalogBaseDocumentModel):
     internal_graph: Any = Field(allow_mutation=False, alias="graph")
     _graph: Optional[QueryGraph] = PrivateAttr(default=None)
 
-    # list of IDs attached to this feature
+    # list of IDs attached to this feature or target
     entity_ids: List[PydanticObjectId] = Field(allow_mutation=False, default_factory=list)
     table_ids: List[PydanticObjectId] = Field(allow_mutation=False, default_factory=list)
     primary_table_ids: List[PydanticObjectId] = Field(allow_mutation=False, default_factory=list)
-    feature_namespace_id: PydanticObjectId = Field(allow_mutation=False, default_factory=ObjectId)
-    feature_list_ids: List[PydanticObjectId] = Field(allow_mutation=False, default_factory=list)
-    deployed_feature_list_ids: List[PydanticObjectId] = Field(
-        allow_mutation=False, default_factory=list
-    )
     user_defined_function_ids: List[PydanticObjectId] = Field(
         allow_mutation=False, default_factory=list
     )
@@ -246,7 +79,7 @@ class FeatureModel(FeatureByteCatalogBaseDocumentModel):
             node = graph.get_node_by_name(node_name)
             op_struct = graph.extract_operation_structure(node=node)
             if len(op_struct.aggregations) != 1:
-                raise ValueError("Feature graph must have exactly one aggregation output")
+                raise ValueError("Feature or target graph must have exactly one aggregation output")
 
             values["dtype"] = op_struct.aggregations[0].dtype
         return values
@@ -308,7 +141,7 @@ class FeatureModel(FeatureByteCatalogBaseDocumentModel):
 
     def extract_operation_structure(self) -> GroupOperationStructure:
         """
-        Extract feature operation structure based on query graph.
+        Extract feature or target operation structure based on query graph.
 
         Returns
         -------
@@ -318,7 +151,60 @@ class FeatureModel(FeatureByteCatalogBaseDocumentModel):
         operation_structure = self.graph.extract_operation_structure(self.node)
         return operation_structure.to_group_operation_structure()
 
-    class Settings(FeatureByteCatalogBaseDocumentModel.Settings):
+
+class FeatureModel(BaseFeatureTargetModel):
+    """
+    Model for Feature asset
+
+    id: PydanticObjectId
+        Feature id of the object
+    name: str
+        Feature name
+    dtype: DBVarType
+        Variable type of the feature
+    graph: QueryGraph
+        Graph contains steps of transformation to generate the feature
+    node_name: str
+        Node name of the graph which represent the feature
+    tabular_source: TabularSource
+        Tabular source used to construct this feature
+    readiness: FeatureReadiness
+        Feature readiness
+    version: VersionIdentifier
+        Feature version
+    online_enabled: bool
+        Whether to make this feature version online enabled
+    definition: str
+        Feature definition
+    entity_ids: List[PydanticObjectId]
+        Entity IDs used by the feature
+    table_ids: List[PydanticObjectId]
+        Table IDs used by the feature
+    primary_table_ids: Optional[List[PydanticObjectId]]
+        Primary table IDs of the feature (auto-derive from graph)
+    feature_namespace_id: PydanticObjectId
+        Feature namespace id of the object
+    feature_list_ids: List[PydanticObjectId]
+        FeatureList versions which use this feature version
+    deployed_feature_list_ids: List[PydanticObjectId]
+        Deployed FeatureList versions which use this feature version
+    created_at: Optional[datetime]
+        Datetime when the Feature was first saved
+    updated_at: Optional[datetime]
+        When the Feature get updated
+    """
+
+    readiness: FeatureReadiness = Field(allow_mutation=False, default=FeatureReadiness.DRAFT)
+    online_enabled: bool = Field(allow_mutation=False, default=False)
+
+    # list of IDs attached to this feature
+    feature_namespace_id: PydanticObjectId = Field(allow_mutation=False, default_factory=ObjectId)
+    feature_list_ids: List[PydanticObjectId] = Field(allow_mutation=False, default_factory=list)
+    deployed_feature_list_ids: List[PydanticObjectId] = Field(
+        allow_mutation=False, default_factory=list
+    )
+
+    class Settings(BaseFeatureTargetModel.Settings):
         """
         MongoDB settings
         """
@@ -355,20 +241,3 @@ class FeatureModel(FeatureByteCatalogBaseDocumentModel):
                 ],
             ),
         ]
-
-
-class FeatureSignature(FeatureByteBaseModel):
-    """
-    FeatureSignature class used in FeatureList object
-
-    id: PydanticObjectId
-        Feature id of the object
-    name: str
-        Name of the feature
-    version: VersionIdentifier
-        Feature version
-    """
-
-    id: PydanticObjectId
-    name: Optional[StrictStr]
-    version: VersionIdentifier
