+++ featurebyte/schema/feature_list_namespace.py
@@ -4,12 +4,12 @@ FeatureListNamespace API payload scheme
 from typing import List, Optional
 
 from featurebyte.models.base import FeatureByteBaseModel, PydanticObjectId
-from featurebyte.models.feature import DefaultVersionMode
 from featurebyte.models.feature_list import (
     FeatureListNamespaceModel,
     FeatureListStatus,
     FeatureReadinessDistribution,
 )
+from featurebyte.models.feature_namespace import DefaultVersionMode
 from featurebyte.schema.common.base import BaseDocumentServiceUpdateSchema, PaginationMixin
