+++ featurebyte/service/online_enable.py
@@ -8,8 +8,9 @@ from typing import Any, Optional
 from bson.objectid import ObjectId
 
 from featurebyte.feature_manager.model import ExtendedFeatureModel
-from featurebyte.models.feature import FeatureModel, FeatureNamespaceModel
+from featurebyte.models.feature import FeatureModel
 from featurebyte.models.feature_list import FeatureListModel
+from featurebyte.models.feature_namespace import FeatureNamespaceModel
 from featurebyte.models.online_store import OnlineFeatureSpec
 from featurebyte.persistent import Persistent
 from featurebyte.schema.feature import FeatureServiceUpdate
