+++ tests/unit/conftest.py
@@ -37,8 +37,8 @@ from featurebyte.enum import AggFunc, InternalName
 from featurebyte.feature_manager.model import ExtendedFeatureListModel
 from featurebyte.models.base import DEFAULT_CATALOG_ID, VersionIdentifier
 from featurebyte.models.credential import CredentialModel
-from featurebyte.models.feature import FeatureReadiness
 from featurebyte.models.feature_list import FeatureListStatus
+from featurebyte.models.feature_namespace import FeatureReadiness
 from featurebyte.models.task import Task as TaskModel
 from featurebyte.models.tile import TileSpec
 from featurebyte.query_graph.graph import GlobalQueryGraph
